Refactor structure of components.js
