+++ frontend/components/app.js
@@ -1,0 +1,17 @@
+const {h} = require('preact')
+const connect = require('unistore/preact').connect
+
+const TopLevelRouter = require('./router').TopLevelRouter
+const Footer = require('./common/footer')
+const LoadingOverlay = require('./common/loadingOverlay')
+const NavbarAuth = require('./common/navbarAuth')
+const NavbarUnauth = require('./common/navbarUnauth')
+
+const Navbar = connect((state) => ({
+  walletIsLoaded: state.walletIsLoaded,
+}))(({walletIsLoaded}) => (walletIsLoaded ? h(NavbarAuth) : h(NavbarUnauth)))
+
+const App = () =>
+  h('div', {class: 'wrap'}, h(LoadingOverlay), h(Navbar), h(TopLevelRouter), h(Footer))
+
+module.exports = App
