+++ frontend/components/pages/dashboard/balance.js
@@ -1,0 +1,17 @@
+const {h} = require('preact')
+const printAda = require('../../../printAda')
+
+const Balance = ({balance}) =>
+  h(
+    'div',
+    {class: 'balance-block'},
+    h('h2', undefined, 'Balance'),
+    h(
+      'p',
+      {class: 'balance-value'},
+      h('span', undefined, isNaN(Number(balance)) ? balance : `${printAda(balance)}`),
+      h('img', {class: 'ada-sign-big', alt: 'ADA', src: '/assets/ada.png'})
+    )
+  )
+
+module.exports = Balance
